+++ swagger-request-validator-springmvc/src/test/java/com/atlassian/oai/validator/springmvc/example/simple/RestServiceApplication.java
@@ -1,14 +1,14 @@
 //CHECKSTYLE:OFF // HideUtilityClassConstructor: this is no utility class
-package com.atlassian.oai.validator.springmvc.example;
+package com.atlassian.oai.validator.springmvc.example.simple;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class SwaggerRequestValidationApplication {
+public class RestServiceApplication {
 
     public static void main(final String[] args) {
-        SpringApplication.run(SwaggerRequestValidationApplication.class, args);
+        SpringApplication.run(RestServiceApplication.class, args);
     }
 }
 //CHECKSTYLE:ON